requirements.txt
absl-py==1.3.0
allennlp==2.10.1
altair==4.2.0
astor==0.8.1
asttokens==2.2.1
attrdict==2.0.1
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
base58==2.1.1
bce-python-sdk==0.8.74
beautifulsoup4==4.11.1
blinker==1.5
blis==0.7.9
boto3==1.26.28
botocore==1.29.28
cached-path==1.1.6
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
contourpy==1.0.6
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
cymem==2.0.7
Cython==0.29.32
decorator==5.1.1
dill==0.3.6
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.2.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.0.4
executing==1.2.0
fairscale==0.4.6
filelock==3.7.1
Flask==2.2.2
Flask-Babel==2.0.0
fonttools==4.38.0
future==0.18.2
gdown==4.4.0
gitdb==4.0.10
GitPython==3.1.29
google-api-core==2.11.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
h5py==3.7.0
huggingface-hub==0.10.1
idna==3.4
imageio==2.22.4
imgaug==0.4.0
importlib-metadata==5.1.0
iniconfig==1.1.1
ipython==8.7.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonnet==0.19.1
jsonschema==4.17.3
kiwisolver==1.4.4
langcodes==3.3.0
lanms-neo==1.0.2
lmdb==1.4.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
more-itertools==9.0.0
multiprocess==0.70.14
murmurhash==1.0.9
networkx==2.8.8
nltk==3.8
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
overrides==7.3.1
packaging==22.0
paddle-bfloat==0.1.7
paddlepaddle==2.4.1
pandas==1.5.2
parso==0.8.3
pathtools==0.1.2
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pipreqs==0.4.11
pluggy==1.0.0
Polygon3==3.0.9.1
prefetch-generator==1.0.1
premailer==3.10.0
preshed==3.0.8
promise==2.3
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycryptodome==3.16.0
pydantic==1.8.2
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
PyMuPDF==1.20.2
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
rapidfuzz==2.13.6
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
s3transfer==0.6.0
sacremoses==0.0.53
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.12.1
semver==2.13.0
sentencepiece==0.1.97
sentry-sdk==1.11.1
setproctitle==1.3.2
shapely==2.0.0
shortuuid==1.0.11
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
soupsieve==2.3.2.post1
spacy==3.3.1
spacy-legacy==3.0.10
spacy-loggers==1.0.4
srsly==2.4.5
stack-data==0.6.2
streamlit==1.15.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
termcolor==1.1.0
thinc==8.0.17
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2022.10.10
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.12.0+cu116
torchaudio==0.12.0+cu116
torchtext==0.6.0
torchvision==0.13.0+cu116
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
transformers==4.20.1
typer==0.4.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
vietocr==0.3.11
visualdl==2.4.1
wandb==0.12.21
wasabi==0.10.1
watchdog==2.2.0
wcwidth==0.2.5
Werkzeug==2.2.2
yarg==0.1.9
zipp==3.11.0
